backEnd/requirements.txt
fastapi
uvicorn[standard]
psycopg2
bcrypt
PyJWT